% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{candEdgeSpots}
\alias{candEdgeSpots}
\title{candEdgeSpots}
\usage{
candEdgeSpots(elu.dist.sig, dist.mat)
}
\arguments{
\item{elu.dist.sig}{Distance matrix between significant spots and entire spots.}

\item{dist.mat}{Distance matrix of all spots.}
}
\value{
A character vector of candidate edge spots.
}
\description{
Determine the edge spots of each module.
}
