% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combinePvalsByStouffer}
\alias{combinePvalsByStouffer}
\title{combinePvalsByStouffer}
\usage{
combinePvalsByStouffer(obj.st.lst)
}
\arguments{
\item{obj.st.lst}{A list of spatial Seurat objects containing spatial data.}
}
\value{
Combined P values with adjusted FDR values.
}
\description{
Combine P values using the Stouffer's method.
}
