% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distScaleToOne}
\alias{distScaleToOne}
\title{distScaleToOne}
\usage{
distScaleToOne(obj.st.lst)
}
\arguments{
\item{obj.st.lst}{A list of spatial Seurat objects.}
}
\value{
A list of scaled distance matrices for each object in obj.st.lst.
}
\description{
Scale distances within each module to a range of \link{0, 1}.
}
