% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{fitDistLinesByGlm}
\alias{fitDistLinesByGlm}
\title{fitDistLinesByGlm}
\usage{
fitDistLinesByGlm(obj.st.lst, plot.tar, degree = 4)
}
\arguments{
\item{obj.st.lst}{A list of spatial seurat objects.}

\item{plot.tar}{Vector of gene names or metadata column names to plot.}

\item{degree}{Integer specifying the degree of splines for smoothing.}
}
\value{
ggplot object.
}
\description{
Fits generalized linear models (GLMs) to gene expression data against spatial distance and visualizes the fitted curves with confidence intervals.
}
