% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{fitDistLinesByWindows}
\alias{fitDistLinesByWindows}
\title{fitDistLinesByWindows}
\usage{
fitDistLinesByWindows(obj.st.lst, plot.tar, win = 10)
}
\arguments{
\item{obj.st.lst}{A list of spatial seurat objects.}

\item{plot.tar}{Vector of gene names or metadata column names to plot.}

\item{win}{Integer specifying the window size for averaging data points.}
}
\value{
ggplot object.
}
\description{
Curves of gene expression with spatial distance.
}
