% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitNorm}
\alias{fitNorm}
\title{fitNorm}
\usage{
fitNorm(obj.st.lst)
}
\arguments{
\item{obj.st.lst}{A list of spatial seurat data objects.}
}
\value{
est.params Vector of estimated parameters from the normal distribution fitting.
}
\description{
Fit medulla score distribution and corresponding parameters.
}
