% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glmFitTest}
\alias{glmFitTest}
\title{glmFitTest}
\usage{
glmFitTest(
  obj.st.lst,
  tar.genes = NULL,
  other.meta = NULL,
  query.name = "Distance",
  ncores = 8
)
}
\arguments{
\item{obj.st.lst}{A list of spatial Seurat objects containing spatial data.}

\item{tar.genes}{Target genes to be tested for association with distance.}

\item{other.meta}{Other metadata column names to be tested for association with distance.}

\item{query.name}{Column name which contains the distance metric, default: "Distance".}

\item{ncores}{Number of cores to use for parallel processing, default: 8.}
}
\value{
A list of results (sig.df) containing coefficients from the generalized linear models.
}
\description{
Inferring the gene or cell type associated with distance significance using generalized linear models.
}
