% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{medullaScore}
\alias{medullaScore}
\title{medullaScore}
\usage{
medullaScore(obj.st.lst, medulla.genes)
}
\arguments{
\item{obj.st.lst}{A list of spatial seurat data objects.}

\item{medulla.genes}{A vector of genes associated with the medulla.}
}
\value{
Updated obj.st.lst with added medulla scores.
}
\description{
Calculates medulla scores for each spot in spatial seurat data objects using Seurat's method.
}
