% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeEdgeSpotsWithOutInternalSpots}
\alias{removeEdgeSpotsWithOutInternalSpots}
\title{removeEdgeSpotsWithOutInternalSpots}
\usage{
removeEdgeSpotsWithOutInternalSpots(edge.dist, labels)
}
\arguments{
\item{edge.dist}{Edge distance matrix.}

\item{labels}{A vector of labels of spots.}
}
\value{
A vector of removed spots.
}
\description{
Remove low confidence edge spots based on edge distance matrix.
}
