% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeLowConfModules}
\alias{removeLowConfModules}
\title{removeLowConfModules}
\usage{
removeLowConfModules(elu.dist.sig, module.size = 10)
}
\arguments{
\item{elu.dist.sig}{Distance matrix between significant spots and other spots.}

\item{module.size}{Minimum module size for a confident module, default: 10.}
}
\value{
A list containing removed and remaining spots.
}
\description{
Identify low confidence modules of spots using traversal strategy.
}
