% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{signifTestByZtest}
\alias{signifTestByZtest}
\title{signifTestByZtest}
\usage{
signifTestByZtest(obj, est.params)
}
\arguments{
\item{obj}{Spatial seurat object data.}

\item{est.params}{A vector of estimated parameters of normal distribution (mean and standard deviation).}
}
\value{
Updated Seurat object with adjusted p-values.
}
\description{
Z-test to identify spots with significant Medulla scores.
}
