% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsoHis.R
\name{tsoHis}
\alias{tsoHis}
\title{Find medulla clusters and corresponding edges.}
\usage{
tsoHis(
  obj.st.lst,
  medulla.genes = NULL,
  call.xgb = FALSE,
  p.cut = 1e-05,
  module.size = 10,
  remove.spots = NULL
)
}
\arguments{
\item{obj.st.lst}{A list of thymus spatial seurat objects.}

\item{medulla.genes}{A vector of genes associated with the medulla.}

\item{call.xgb}{Call trained XGBoost model or not. Default: FALSE.}

\item{p.cut}{Cutoff p-value to determine significant spots. Default is 1e-5.}

\item{module.size}{Minimum module size for modules. Default is 10.}

\item{remove.spots}{Optional list of spots to be removed from analysis.}

\item{out.figs.dir}{Directory to save output figures. If provided, a density plot of medulla scores will be saved.}
}
\value{
Updated obj.st.lst with identified medulla clusters and edges.
}
\description{
This function identifies medulla and cortex spots and their corresponding boundaries based on thymus spatial transcriptomics (ST) data.
}
\examples{
# Example usage:
sp.obj <- system.file("data/thymus_T2.RDS", package = "thymusTSO") \%>\% readRDS()
sp.obj <- tsoHis(sp.obj)
}
