% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unpdateSigSpots}
\alias{unpdateSigSpots}
\title{unpdateSigSpots}
\usage{
unpdateSigSpots(elu.dist, sig.spots)
}
\arguments{
\item{elu.dist}{Eulerian distance matrix between spots.}

\item{sig.spots}{Current significance spots identified.}
}
\value{
Updated vector of significance spots.
}
\description{
Update significance spots based on Eulerian distance matrix.
}
