% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateModuleCenter}
\alias{updateModuleCenter}
\title{updateModuleCenter}
\usage{
updateModuleCenter(obj, modules, edge.spots)
}
\arguments{
\item{obj}{Spatial Seurat object data containing spatial and image coordinates.}

\item{modules}{List of remained significant modules identified.}

\item{edge.spots}{List of edge spots to exclude from center calculation.}
}
\value{
Updated modules with updated center spots.
}
\description{
Update center spot of each remained module.
}
